+++ CHANGELOG.md
@@ -3,9 +3,7 @@
 * Add deprecation warning around client machines and how they are handled in 8.0.
 * Handle new AMS stack conventions and dashboards; auto-discover all available dashboards and folders.
 * Add options for specifying replacement or additive dashboard lists to AMS stack.
-* TODO GCP move some values from labels to custom metadata
-* TODO GCP support setting custom metadata values
-* TODO Add option to map and expose ports in Docker backend on a 1:1 pairing (eg 3100 node 1, 3101 node 2 etc) - so not workarounds are needed to access Aerospike clusters on Docker Desktop from the Desktop.
+* Add option to map and expose ports in Docker backend on a 1:1 pairing (eg 3100 node 1, 3101 node 2 etc) - so not workarounds are needed to access Aerospike clusters on Docker Desktop from the Desktop.
 * TODO Add MacOS packaging and signing to makefile to move fully away from bash scripts.
 * TODO Add option for auto-expiring clusters in GCP and AWS.
 * TODO AeroLab Support for deploying Amazon 2023 server >= 6.4.